Data.Universe.Instances.Ord

Documentation

An Ord instance for functions, given the input is Finite and the output is Ord. Compares pointwise, with higher priority to inputs that appear earlier in universeF.

class Eq a => Ord a where

The Ord class is used for totally ordered datatypes.

Instances of Ord can be derived for any user-defined datatype whose constituent types are in Ord. The declared order of the constructors in the data declaration determines the ordering in derived Ord instances. The Ordering datatype allows a single comparison to determine the precise ordering of two objects.

The Haskell Report defines no laws for Ord. However, <= is customarily expected to implement a non-strict partial order and have the following properties:

Transitivity

if x <= y && y <= z = True, then x <= z = True

Reflexivity

x <= x = True

Antisymmetry

if x <= y && y <= x = True, then x == y = True

Note that the following operator interactions are expected to hold:

x >= y = y <= x

x < y = x <= y && x /= y

x > y = y < x

x < y = compare x y == LT

x > y = compare x y == GT

x == y = compare x y == EQ

min x y == if x <= y then x else y = True

max x y == if x >= y then x else y = True

Note that (7.) and (8.) do not require min and max to return either of their arguments. The result is merely required to equal one of the arguments in terms of (==).

Minimal complete definition: either compare or <=. Using compare can be more efficient for complex types.

Minimal complete definition

compare | (<=)

Methods

compare :: a -> a -> Ordering

(<) :: a -> a -> Bool infix 4

(<=) :: a -> a -> Bool infix 4

(>) :: a -> a -> Bool infix 4

(>=) :: a -> a -> Bool infix 4

max :: a -> a -> a

min :: a -> a -> a

Orphan instances

(Finite a, Ord a) => Ord (Endo a)

Instance details

Methods

compare :: Endo a -> Endo a -> Ordering

(<) :: Endo a -> Endo a -> Bool

(<=) :: Endo a -> Endo a -> Bool

(>) :: Endo a -> Endo a -> Bool

(>=) :: Endo a -> Endo a -> Bool

max :: Endo a -> Endo a -> Endo a

min :: Endo a -> Endo a -> Endo a

(Finite a, Ord b) => Ord (a -> b)

Instance details

Methods

compare :: (a -> b) -> (a -> b) -> Ordering

(<) :: (a -> b) -> (a -> b) -> Bool

(<=) :: (a -> b) -> (a -> b) -> Bool

(>) :: (a -> b) -> (a -> b) -> Bool

(>=) :: (a -> b) -> (a -> b) -> Bool

max :: (a -> b) -> (a -> b) -> a -> b

min :: (a -> b) -> (a -> b) -> a -> b