Data.Universe.Instances.Show

Description

A Show instance for functions, given the input is Finite and both the input and output are Show.

Orphan instances

(Finite a, Show a) => Show (Endo a)

Instance details

Methods

showsPrec :: Int -> Endo a -> ShowS

show :: Endo a -> String

showList :: [Endo a] -> ShowS

(Finite a, Show a, Show b) => Show (a -> b)

Instance details

Methods

showsPrec :: Int -> (a -> b) -> ShowS

show :: (a -> b) -> String

showList :: [a -> b] -> ShowS